requirements.txt
# Core ML Libraries
# insightface==0.7.3  # Skip for now due to build issues
onnxruntime-gpu==1.23.0
onnxruntime==1.23.0
mtcnn==1.0.0
tensorflow>=2.13.0
opencv-python>=4.8.0
numpy>=1.26.0
scikit-learn>=1.3.0

# Vector Database
qdrant-client==1.7.0

# API Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# Image Processing & Storage
Pillow==10.2.0
python-magic==0.4.27
boto3==1.34.20

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
aiofiles==23.2.1

# Testing
pytest==7.4.3
pytest-asyncio==0.23.3
httpx==0.26.0
pytest-cov==4.1.0
